% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2mat.R
\name{cor2mat}
\alias{cor2mat}
\title{Correlations to matrix}
\usage{
cor2mat(samplestats)
}
\arguments{
\item{samplestats}{sample statistics from an TIRT model (Mplus output)}
}
\value{
correlation matrix of tetrachoric correlation
}
\description{
Correlations to matrix
}
