% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctRmsea.R
\name{correctRmsea}
\alias{correctRmsea}
\title{correctRmsea}
\usage{
correctRmsea(n, p, chisq, df, N)
}
\arguments{
\item{n}{block size}

\item{p}{no of blocks in questionnaire}

\item{chisq}{chi square value}

\item{df}{degrees of freedom}

\item{N}{sample size}
}
\value{
cRmsea
}
\description{
In questionnaires with block size n > 2, the degrees of freedom need to
be corrected and the RMSEA value need to be
recalculated (Brown & Maydeu-Olivares, 2012).
This function can be used to recalculate the RMSEA value of a TIRT model.
}
