% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designLoadings.R
\name{designLoadings}
\alias{designLoadings}
\title{Design Loadings}
\usage{
designLoadings(all.items, quest, no.traits)
}
\arguments{
\item{all.items}{item coding table (item-trait correspondence, coding of item, ...)}

\item{quest}{quest which questionnaire "BT"/"SD"/"BI"/"OS"... (name in all.items)}

\item{no.traits}{number of traits in the questionnaire}
}
\value{
matrix of design loadings of a questionnaire of interest
}
\description{
Generates a matrix which contains information about which trait is measured by which
item and is it positively or negatively worded.
}
