% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designMatrix.R
\name{designMatrix}
\alias{designMatrix}
\title{Design matrix}
\usage{
designMatrix(no.b)
}
\arguments{
\item{no.b}{number of blocks in the questionnaire}
}
\value{
matrix with loadings per design
}
\description{
In the TIRT model some factor loadings are negative per design. This function
creates a matrix for a questionnaire with no.b blocks which contains the loadings
which a set per design of the TIRT model.
}
