% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load2mat.R
\name{load2mat}
\alias{load2mat}
\title{load2mat}
\usage{
load2mat(mplus.pars, design.load, design.mat)
}
\arguments{
\item{mplus.pars}{Mplus parameters, read in with Mplus Automation}

\item{design.load}{loading matrix, rows = items, columns = traits}

\item{design.mat}{MFC design matrix: rows = pairwise comparisons, columns = items}
}
\value{
loading matrix with Mplus estimates: rows = items, columns = traits
}
\description{
Mplus loadings to matrix
}
