% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reQuads.R
\name{reQuads}
\alias{reQuads}
\title{Recode quads: ranks to binary outcomes (partial ranking)}
\usage{
reQuads(df, fbname)
}
\arguments{
\item{df}{data frame with raw data}

\item{fbname}{name of the questionnaire}
}
\value{
data frame with binary coded data
}
\description{
Recode quads: ranks to binary outcomes (partial ranking)
}
