% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh2vec.R
\name{thresh2vec}
\alias{thresh2vec}
\title{thresh2vec}
\usage{
thresh2vec(samplestats)
}
\arguments{
\item{samplestats}{sample statistics from an TIRT model (Mplus output)}
}
\description{
TIRT-thresholds from Mplus output to vector
}
