% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tirtMplusSyntax.R
\name{tirtMplusSyntax}
\alias{tirtMplusSyntax}
\title{tirtMplusSyntax}
\usage{
tirtMplusSyntax(
  design.load,
  names.pairs,
  item.short,
  id.var,
  file.inp = NULL,
  file.data,
  title,
  out.command,
  fscores.file,
  missings.as
)
}
\arguments{
\item{design.load}{loading matrix of items, rows=items, cols=traits}

\item{names.pairs}{if NULL they will be named}

\item{item.short}{shortening to indicate from which questionnaire the items belong}

\item{id.var}{ID-variable or case-number in quotes}

\item{file.inp}{file name (directory and name) for Mplus syntax (saved as R-object)}

\item{file.data}{file name for the data file (.R) without directory: assumed in the same folder as file.inp}

\item{title}{title for the model}

\item{out.command}{character string passed to the output command (for example "sampstat standardized;")}

\item{fscores.file}{name of the file where (and with what name) to save the factor-scores  (e.g., "fscores.dat")}

\item{missings.as}{how missings are coded in the dataset (e.g., "-99")}
}
\value{
Mplus syntax as list of character vectors, commands as list entries
syntax written to file.inp
syntax printed to console (with cat)
create syntax: syntax <- tirt.mplus.syntax(...)
subsequently: cat(paste(syntax, collapse="\n\n"), file="mplus.inp")
}
\description{
Function to create Mplus syntax for TIRT (only for blocks of 3 items, full ranking,
assumes items are ordered blockwise)
}
