% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retri.R
\name{retri}
\alias{retri}
\title{Recode triplets: ranks to binary outcomes}
\usage{
retri(df, fbname)
}
\arguments{
\item{df}{data frame with raw data}

\item{fbname}{name of the questionnaire}
}
\value{
data frame with binary coded data
}
\description{
Recode triplets: ranks to binary outcomes
}
